CLDR-16369 Move VoteStatus enum from VettingViewer to VoteResolver
-VettingViewer should depend on VoteResolver, not vice-versa
